requirements.txt
black==23.9.0
click==8.1.7
pytest==7.4.2
pytest-cov==4.1.0
pylint==2.17.5
nbval==0.10.0

pandas
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-secret-manager
oauth2client